+++ example/lib/main.dart
@@ -58,7 +58,7 @@ class _MyAppState extends State<MyApp> {
                               ? WaveformRecorder(
                                   height: 48,
                                   controller: _waveController,
-                                  onEndRecording: _onEndRecording,
+                                  onRecordingStopped: _onRecordingStopped,
                                 )
                               : TextField(
                                   controller: _textController,
@@ -93,7 +93,7 @@ class _MyAppState extends State<MyApp> {
         false => _waveController.startRecording(),
       };
 
-  void _onEndRecording() => _textController.text = ''
+  void _onRecordingStopped() => _textController.text = ''
       '${_waveController.url}: '
       '${_waveController.length.inMilliseconds / 1000} seconds';