+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+## 1.2
+
+* change onStartRecording => onRecordingStarted
+* change onEndRecording => onRecordingStopped
+* fixed issue on second recording -- yay asserts!
+
 ## 1.0.1
 
 * fix missing file.
-
 
 ## 1.0.0